[CLEANUP] Refactorings
