+++ Classes/Exception/DataException.php
@@ -29,6 +29,6 @@ namespace Ipf\Bib\Exception;
 /**
  * Exception when reading or writing data fails
  */
-class DataException extends \Exception{
+class DataException extends \Exception {
 
 }
