+++ Configuration/TCA/Authorships.php
