+++ Resources/Private/Templates/List/Blocks.html
@@ -1,4 +1,3 @@
-
 <h1>Blocks</h1>
 
 <!-- ###SPACER_BLOCK### start -->
