+++ Resources/Private/Templates/List/Items.html
@@ -1,6 +1,3 @@
-
-
-
 <h1>Single item block</h1>
 
 
@@ -142,7 +139,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -183,7 +180,7 @@
 
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -217,7 +214,7 @@
 <!-- ###HAS_HOWPUBLISHED### end -->
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -241,13 +238,13 @@
 
 <div class="tx_bib-block">
 <!-- ###HAS_EDITOR### begin -->In ###EDITOR###, ###LABEL_EDITOR###,<!-- ###HAS_EDITOR### end -->
-<!-- ###HAS_BOOKTITLE### begin -->###BOOKTITLE###<!-- 
-	###HAS_BOOKTITLE### end --><!-- ###HAS_VOLUME### begin 
+<!-- ###HAS_BOOKTITLE### begin -->###BOOKTITLE###<!--
+	###HAS_BOOKTITLE### end --><!-- ###HAS_VOLUME### begin
 --> ###LABEL_VOLUME### ###VOLUME###<!-- ###HAS_VOLUME### end -->
 
 <!-- ###HAS_SERIES### begin -->
 ###LABEL_OF_SERIES### <span class="tx_bib-series">###SERIES###</span>
-<!-- ###HAS_SERIES### end --><!-- 
+<!-- ###HAS_SERIES### end --><!--
 ###HAS_PAGES### begin -->, ###LABEL_PAGE### ###PAGES###.<!-- ###HAS_PAGES### end -->
 </div>
 
@@ -256,7 +253,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -283,8 +280,8 @@
 <!-- ###INCOLLECTION_DATA### begin -->
 <div class="tx_bib-block">
 <!-- ###HAS_EDITOR### begin -->In ###EDITOR###, ###LABEL_EDITOR###, <!-- ###HAS_EDITOR### end -->
-<!-- ###HAS_BOOKTITLE### begin -->###BOOKTITLE###<!-- 
-	###HAS_BOOKTITLE### end --><!-- 
+<!-- ###HAS_BOOKTITLE### begin -->###BOOKTITLE###<!--
+	###HAS_BOOKTITLE### end --><!--
 	###HAS_VOLUME### begin --> ###LABEL_VOLUME### ###VOLUME###<!-- ###HAS_VOLUME### end -->
 <!-- ###HAS_SERIES### begin -->
 ###LABEL_OF_SERIES### <span class="tx_bib-series">###SERIES###</span>
@@ -306,7 +303,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -340,7 +337,7 @@
 <!-- ###HAS_EVENT_NAME### end -->
 
 <div class="tx_bib-block">
-###EVENT_PLACE###<!-- 
+###EVENT_PLACE###<!--
 ###HAS_EVENT_DATE### begin -->, ###EVENT_DATE###<!-- ###HAS_EVENT_DATE### end -->
 </div>
 
@@ -349,7 +346,7 @@
 <!-- ###HAS_ORGANIZATION### end -->
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -390,7 +387,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -420,7 +417,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_ORGANIZATION### begin -->###ORGANIZATION###, <!-- ###HAS_ORGANIZATION### end --><!-- 
+<!-- ###HAS_ORGANIZATION### begin -->###ORGANIZATION###, <!-- ###HAS_ORGANIZATION### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -451,7 +448,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_SCHOOL### begin -->###SCHOOL###, <!-- ###HAS_SCHOOL### end --><!-- 
+<!-- ###HAS_SCHOOL### begin -->###SCHOOL###, <!-- ###HAS_SCHOOL### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -474,7 +471,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_SCHOOL### begin -->###SCHOOL###, <!-- ###HAS_SCHOOL### end --><!-- 
+<!-- ###HAS_SCHOOL### begin -->###SCHOOL###, <!-- ###HAS_SCHOOL### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -508,7 +505,7 @@
 <!-- ###HAS_AFFILIATION### end -->
 
 <div class="tx_bib-block">
-<!-- ###HAS_INSTITUTION### begin -->###INSTITUTION###, <!-- ###HAS_INSTITUTION### end --><!-- 
+<!-- ###HAS_INSTITUTION### begin -->###INSTITUTION###, <!-- ###HAS_INSTITUTION### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -587,7 +584,7 @@
 </div>
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -614,7 +611,7 @@
 <!-- ###HAS_AFFILIATION### end -->
 
 <div class="tx_bib-block">
-<!-- ###HAS_SCHOOL### begin -->###SCHOOL###, <!-- ###HAS_SCHOOL### end --><!-- 
+<!-- ###HAS_SCHOOL### begin -->###SCHOOL###, <!-- ###HAS_SCHOOL### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -665,7 +662,7 @@
 <!-- ###HAS_INSTITUTE### end -->
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
 
@@ -700,7 +697,7 @@
 <!-- ###HAS_INSTITUTE### end -->
 
 <div class="tx_bib-block">
-<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!-- 
+<!-- ###HAS_PUBLISHER### begin -->###LABEL_PUBLISHER### ###PUBLISHER###, <!-- ###HAS_PUBLISHER### end --><!--
 ###HAS_ADDRESS### begin -->###ADDRESS###<!-- ###HAS_ADDRESS### end -->
 </div>
