+++ Resources/Private/Templates/main.html
@@ -1,4 +1,3 @@
-
 <h1>The list view main template</h1>
 
 <!-- ###LIST_VIEW### begin -->
