+++ ext_tables.php
@@ -18,8 +18,8 @@ $TCA['tx_bib_domain_model_reference'] = [
 		'enablecolumns' => [
 			'disabled' => 'hidden',
 		],
-		'dynamicConfigFile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath($_EXTKEY) . 'Configuration/Tca/References.php',
-		'iconfile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extRelPath($_EXTKEY) . 'Resources/Public/Icons/icon_tx_bib_domain_model_reference.png',
+		'dynamicConfigFile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath('bib') . 'Configuration/TCA/References.php',
+		'iconfile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extRelPath('bib') . 'Resources/Public/Icons/icon_tx_bib_domain_model_reference.png',
 	],
 	'feInterface' => [
 		'fe_admin_fieldList' => 'hidden,bibtype,citeid,title,journal,year,month,day,volume,number,number2,pages,abstract,affiliation,note,annotation,keywords,tags,file_url,web_url,web_url_date,misc, editor,publisher,address,howpublished,series,edition,chapter,booktitle,school,institute,organization,institution,event_name,event_place,event_date,state,type,ISBN,ISSN,DOI,extern,reviewed,in_library,borrowed_by',
@@ -38,8 +38,8 @@ $TCA['tx_bib_domain_model_author'] = [
 		'cruser_id' => 'cruser_id',
 		'default_sortby' => 'ORDER BY surname',
 		'delete' => 'deleted',
-		'dynamicConfigFile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath($_EXTKEY) . 'Configuration/Tca/Authors.php',
-		'iconfile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extRelPath($_EXTKEY) . 'Resources/Public/Icons/icon_tx_bib_domain_model_reference.png',
+		'dynamicConfigFile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath('bib') . 'Configuration/TCA/Authors.php',
+		'iconfile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extRelPath('bib') . 'Resources/Public/Icons/icon_tx_bib_domain_model_reference.png',
 	],
 	'feInterface' => [
 		'fe_admin_fieldList' => 'surname,forename,url',
@@ -55,8 +55,8 @@ $TCA['tx_bib_domain_model_authorships'] = [
 		'label_alt_force' => 1,
 		'default_sortby' => 'ORDER BY pub_id DESC, sorting ASC',
 		'delete' => 'deleted',
-		'dynamicConfigFile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath($_EXTKEY) . 'Configuration/Tca/Authorships.php',
-		'iconfile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extRelPath($_EXTKEY) . 'Resources/Public/Icons/icon_tx_bib_domain_model_reference.png',
+		'dynamicConfigFile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath('bib') . 'Configuration/TCA/Authorships.php',
+		'iconfile' => \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extRelPath('bib') . 'Resources/Public/Icons/icon_tx_bib_domain_model_reference.png',
 	],
 	'feInterface' => [
 		'fe_admin_fieldList' => 'pub_id,author_id,sorting',
@@ -70,31 +70,31 @@ $TCA['tx_bib_domain_model_authorships'] = [
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::allowTableOnStandardPages('tx_bib_domain_model_authorships');
 
 // Plugin 1: Publication List
-$TCA['tt_content']['types']['list']['subtypes_excludelist'][$_EXTKEY . '_pi1'] = 'layout,select_key';
-$TCA['tt_content']['types']['list']['subtypes_addlist'][$_EXTKEY . '_pi1'] = 'pi_flexform';
+$TCA['tt_content']['types']['list']['subtypes_excludelist']['bib_pi1'] = 'layout,select_key';
+$TCA['tt_content']['types']['list']['subtypes_addlist']['bib_pi1'] = 'pi_flexform';
 
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPlugin(
 	[
 		'LLL:EXT:bib/Resources/Private/Language/locallang_db.xml:tt_content.list_type_pi1',
-		$_EXTKEY . '_pi1'
+		'bib_pi1'
 	],
 	'list_type'
 );
 
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPiFlexFormValue(
-	$_EXTKEY . '_pi1',
+	'bib_pi1',
 	'FILE:EXT:bib/Configuration/FlexForms/flexform_ds.xml'
 );
 
 
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile(
-	$_EXTKEY,
+	'bib',
 	'Configuration/TypoScript/default',
 	'Publication list defaults'
 );
 
 \TYPO3\CMS\Extbase\Utility\ExtensionUtility::registerPlugin(
-		'Ipf.' . $_EXTKEY,
+		'Ipf.bib',
 		'rest',
 		'Bib REST Service'
 );
